+++ src/map_service/path/finder/finder.h
@@ -12,8 +12,8 @@ public:
     PathFinder() = default;
 
     std::vector<size_t> FindPath(const size_t &from, const size_t &to, const size_t &size,
-                                 const std::vector<std::vector<int>> &AdjacencyTable,
-                                 const std::vector<int> &CounterArray);
+                                 const std::vector<std::vector<size_t>> &AdjacencyTable,
+                                 const std::vector<size_t> &CounterArray);
 
     ~PathFinder() = default;
 };
